pyproject.toml
[tool.poetry]
name = "discord-py-masterclass"
version = "0.1.0"
description = ""
authors = ["FallenDeity <61227305+FallenDeity@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {branch = "master", git = "https://github.com/Rapptz/discord.py.git", extras = ["voice"]}
python-dotenv = "^1.0.0"
gpytranslate = "^1.5.1"
tabulate = "^0.9.0"
jishaku = "^2.5.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.1.6"
pre-commit = "^3.3.3"
pyright = "^1.1.324"
mkdocs-material = "^9.2.5"
pygments = "^2.16.1"
mkdocs = "^1.5.2"
mkdocstrings = {extras = ["python"], version = ">=0.19,<0.25"}
Pillow = "^10.0.0"
CairoSVG = "^2.7.1"
pipwin = "^0.5.2"
mkdocs-glightbox = "^0.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.black]
line-length = 120
target-version = ['py39']

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pyright]
pythonVersion = "3.9"
typeCheckingMode = "strict"
reportImportCycles = false
reportPrivateUsage = false
reportPrivateImportUsage = false
reportMissingModuleSource = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
